Move status messages fetching code into internal/repos
