+++ internal/repoupdater/protocol/repoupdater.go
@@ -189,26 +189,3 @@ type ExternalServiceSyncResult struct {
 	ExternalService api.ExternalService
 	Error           string
 }
-
-type CloningProgress struct {
-	Message string
-}
-
-type ExternalServiceSyncError struct {
-	Message           string
-	ExternalServiceId int64
-}
-
-type SyncError struct {
-	Message string
-}
-
-type StatusMessage struct {
-	Cloning                  *CloningProgress          `json:"cloning"`
-	ExternalServiceSyncError *ExternalServiceSyncError `json:"external_service_sync_error"`
-	SyncError                *SyncError                `json:"sync_error"`
-}
-
-type StatusMessagesResponse struct {
-	Messages []StatusMessage `json:"messages"`
-}
